Simplify auth config lookup
